requirements.txt
aiohappyeyeballs
aiohttp
aiosignal
annotated-types
anyio
attrs
blinker
certifi
charset-normalizer
click
colorama
dataclasses-json
distro
dotenv
flask
Flask
frozenlist
greenlet
h11
httpcore
httpx
httpx-sse
idna
itsdangerous
Jinja2
jiter
jsonpatch
jsonpointer
jsonschema
jsonschema-specifications
langchain
langchain-community
langchain-core
langchain-mcp-adapters
langchain-openai
langchain-text-splitters
langgraph
langgraph-checkpoint
langgraph-prebuilt
langgraph-sdk
langsmith
Markdown
MarkupSafe
marshmallow
mcp
multidict
mypy_extensions
numpy
openai
orjson
ormsgpack
packaging
pdfkit
pillow
propcache
pydantic
pydantic-settings
pydantic_core
python-dotenv
python-multipart

PyYAML
redis
referencing
regex
reportlab
requests
requests-toolbelt
rpds-py
sniffio
SQLAlchemy
sse-starlette
starlette
tenacity
tiktoken
tqdm
typing-inspect
typing-inspection
typing_extensions
urllib3
uvicorn
Werkzeug
xxhash
yarl
zstandard
mysql-connector-python
jieba
chromadb

mysql
mysql.connector
gunicorn
pymysql